FASToryEnergyConsumptionDataCollection/FASTory_Data_Collection_New/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
Flask==2.0.2
Flask-MQTT==1.1.1
Flask-SQLAlchemy==2.5.1
fonttools==4.29.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mysql-connector-python==8.0.28
numpy==1.22.1
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.0
Pillow==9.0.1
protobuf==3.19.4
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.31
threadpoolctl==3.1.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
